backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Financial Copilot Backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "supabase>=2.3.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "groq>=0.4.0",
    "edge-tts>=6.1.0",
    "SpeechRecognition>=3.10.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]